+++ docs/CHANGELOG.md
@@ -35,6 +35,7 @@
 - commons: improved and generalized itemsjs-adapter
 - backend: improved export/import-scripts
 - common: added generation of viewer to admin-server
+- common: improved visibility-filter
 
 ### new features
 - viewer: added viewer as separate environment to view static data without backend
@@ -66,6 +67,7 @@
 - backend: improved sql/solr for all entities
 - commons: improved and generalized itemsjs-adapter
 - backend: improved export/import-scripts
+- common: improved visibility-filter
 
 ### bug fixes
 - backend: fixed media-export of dependent records